client-vpn/requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws_cdk.aws_logs
aws_cdk.aws_certificatemanager